projects/django/requirements.txt
djangorestframework
djangorestframework-simplejwt
jsonfield
sendgrid
django-rest-swagger
python-dotenv
django-cors-headers
